complete re-write of getLandmark, lsline is gone for good!
Rewrote the code for getLandmark to make it much more readable. Still needs some comments to be added but overall it should be easier to follow the flow of the code. 

In addition, removed several redundant for-loops and REPLACED LS LINE WITH POLYFIT, so there should be no more graphing issues. 

NOTE: The new function has NOT been fully tested. However the testing thus far has shown identical functionality to the original. So it should be good to use. The new function inputs and outputs function identically to the originals.